server/src/main/java/se/cygni/paintbot/socketserver/security/AuthenticationController.java
package se.cygni.paintbot.socketserver.security;

import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.web.bind.annotation.*;

import javax.security.auth.login.AccountNotFoundException;

@RestController
@RequestMapping("/api/v1")
@RequiredArgsConstructor
public class AuthenticationController {
    private final AuthenticationService authenticationService;

    @PostMapping("/authenticate")
    public ResponseEntity<String> authenticate(@RequestBody AuthenticationModel authenticationModel) {
        try {
            String username = authenticationModel.getUsername();
            String password = authenticationModel.getPassword();
            String token = authenticationService.authenticate(username, password);
                return ResponseEntity.ok(token);
        }
        catch (UsernameNotFoundException e) {
            return ResponseEntity.notFound().build();
        }
        catch (Exception e) {
            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
        }
    }
}
